+++ nengo/networks/integrator.py
@@ -1,9 +1,8 @@
-import warnings
-
 import nengo
+from nengo.exceptions import ObsoleteError
 
 
-def Integrator(recurrent_tau, n_neurons, dimensions, net=None, **kwargs):
+class Integrator(nengo.Network):
     """An ensemble that accumulates input and maintains state.
 
     This is accomplished through scaling the input signal and recurrently
@@ -17,32 +16,30 @@ def Integrator(recurrent_tau, n_neurons, dimensions, net=None, **kwargs):
         Number of neurons in the recurrently connected ensemble.
     dimensions : int
         Dimensionality of the input signal and ensemble.
-    kwargs
-        Keyword arguments passed through to ``nengo.Network``.
-
-    Returns
-    -------
-    net : Network
-        The newly built product network, or the provided ``net``.
+    **kwargs
+        Keyword arguments passed through to ``nengo.Network``
+        like 'label' and 'seed'.
 
     Attributes
     ----------
-    net.ensemble : Ensemble
+    ensemble : Ensemble
         The recurrently connected ensemble.
-    net.input : Node
+    input : Node
         Provides the input signal.
     """
-    if net is None:
-        kwargs.setdefault('label', "Integrator")
-        net = nengo.Network(**kwargs)
-    else:
-        warnings.warn("The 'net' argument is deprecated.", DeprecationWarning)
 
-    with net:
-        net.input = nengo.Node(size_in=dimensions)
-        net.ensemble = nengo.Ensemble(n_neurons, dimensions=dimensions)
-        nengo.Connection(net.ensemble, net.ensemble, synapse=recurrent_tau)
-        nengo.Connection(net.input, net.ensemble,
-                         transform=recurrent_tau, synapse=None)
-    net.output = net.ensemble
-    return net
+    def __init__(self, recurrent_tau, n_neurons, dimensions, **kwargs):
+        if 'net' in kwargs:
+            raise ObsoleteError("The 'net' argument is no longer supported.")
+        kwargs.setdefault('label', "Integrator")
+        super(Integrator, self).__init__(**kwargs)
+
+        with self:
+            self.input = nengo.Node(size_in=dimensions)
+            self.ensemble = nengo.Ensemble(n_neurons, dimensions=dimensions)
+            nengo.Connection(self.ensemble, self.ensemble,
+                             synapse=recurrent_tau)
+            nengo.Connection(self.input, self.ensemble,
+                             transform=recurrent_tau, synapse=None)
+
+        self.output = self.ensemble