+++ nengo/spa/thalamus.py
@@ -7,7 +7,7 @@ from nengo.spa.module import Module
 from nengo.utils.compat import iteritems
 
 
-class Thalamus(Module):
+class Thalamus(nengo.networks.Thalamus, Module):
     """A thalamus, implementing the effects for an associated basal ganglia.
 
     See `.spa.BasalGanglia` for information on the basal ganglia, and
@@ -83,12 +83,17 @@ class Thalamus(Module):
         self.gates = {}     # gating ensembles per action (created as needed)
         self.channels = {}  # channels to pass transformed data between modules
 
-        Module.__init__(self, label, seed, add_to_container)
-        nengo.networks.Thalamus(self.bg.actions.count,
-                                n_neurons_per_ensemble=self.neurons_action,
-                                mutual_inhib=self.mutual_inhibit,
-                                threshold=self.threshold_action,
-                                net=self)
+        Module.__init__(self)
+        nengo.networks.Thalamus.__init__(
+            self,
+            dimensions=self.bg.actions.count,
+            n_neurons_per_ensemble=self.neurons_action,
+            mutual_inhib=self.mutual_inhibit,
+            threshold=self.threshold_action,
+            label=label,
+            seed=seed,
+            add_to_container=add_to_container,
+        )
 
     def on_add(self, spa):
         Module.on_add(self, spa)